Moved Feedback Symbols
